Restoring useNick feature as it was working before.
Details:
This feature has been restored in order to keep taking the options.useNicks into consideration.
Jitsi developers (temporarly?) removed this option, but we highly depend on this because adding the user's own nickname (login name) to the resource part of the jid allows us to identify which track belongs to which VS user (see: jitsi-util.ts#getUserIdFromVideoConference()).

Details about why this option was removed from jitsi-lib:
 * https://community.jitsi.org/t/what-is-the-reason-to-remove-usenicks/19098
 * https://github.com/jitsi/lib-jitsi-meet/pull/921